@@ -6,7 +6,7 @@ description: Demonstrates how to use the agora_rtc_engine plugin.
 publish_to: 'none' # Remove this line if you wish to publish to pub.dev
 
 environment:
-  sdk: ">=2.7.0 <3.0.0"
+  sdk: '>=2.12.0 <3.0.0'
 
 dependencies:
   flutter:
@@ -17,21 +17,21 @@ dependencies:
     #   agora_rtc_engine: ^x.y.z
     # See https://dart.dev/tools/pub/dependencies#version-constraints
     # The example app is bundled with the plugin so we use a path dependency on
-    # the parent directory to use the current plugin's version. 
+    # the parent directory to use the current plugin's version.
     path: ../
 
   # The following adds the Cupertino Icons font to your application.
   # Use with the CupertinoIcons class for iOS style icons.
-  cupertino_icons: ^0.1.3
   speech_recognition:
     path: ../../speech_recognition
+  cupertino_icons: ^1.0.2
 
-  permission_handler: ^5.0.1+1
+  permission_handler: ^6.0.1
 
 dev_dependencies:
+  integration_test: ^1.0.2+2
   flutter_test:
     sdk: flutter
-  e2e: ^0.2.1
   flutter_driver:
     sdk: flutter
 
