@@ -1,39 +1,27 @@
 name: agora_rtc_engine
 description: Flutter plugin for Agora Video SDK.
-version: 3.2.1
-homepage: https://www.agora.io
-repository: https://github.com/AgoraIO/Flutter-SDK
-
+version: 4.0.0-rc.3
+homepage: 'https://www.agora.io'
+repository: 'https://github.com/AgoraIO/Flutter-SDK'
 environment:
-  sdk: ">=2.7.0 <3.0.0"
-  flutter: ">=1.12.13+hotfix.6"
-
+  sdk: '>=2.12.0 <3.0.0'
+  flutter: '>=1.12.13+hotfix.6'
 dependencies:
   flutter:
     sdk: flutter
   flutter_web_plugins:
     sdk: flutter
-  json_annotation: ^3.0.0
-
+  json_annotation: ^4.0.0
 dev_dependencies:
+  integration_test: ^1.0.2+2
   flutter_test:
     sdk: flutter
-  e2e: ^0.2.1
   flutter_driver:
     sdk: flutter
-  pedantic: ^1.8.0
-  build_runner: ^1.7.0
-  json_serializable: ^3.2.0
-
-# For information on the generic Dart part of this file, see the
-# following page: https://dart.dev/tools/pub/pubspec
-
-# The following section is specific to Flutter.
+  pedantic: ^1.11.0
+  build_runner: ^1.11.5
+  json_serializable: ^4.0.2
 flutter:
-  # This section identifies this Flutter project as a plugin project.
-  # The 'pluginClass' and Android 'package' identifiers should not ordinarily
-  # be modified. They are used by the tooling to maintain consistency when
-  # adding or updating assets for this project.
   plugin:
     platforms:
       android:
@@ -41,43 +29,4 @@ flutter:
         pluginClass: AgoraRtcEnginePlugin
       ios:
         pluginClass: AgoraRtcEnginePlugin
-      linux:
-        pluginClass: AgoraRtcEnginePlugin
-      macos:
-        pluginClass: AgoraRtcEnginePlugin
-      windows:
-        pluginClass: AgoraRtcEnginePlugin
-      web:
-        fileName: agora_rtc_engine_web.dart
-        pluginClass: AgoraRtcEngineWeb
 
-  # To add assets to your plugin package, add an assets section, like this:
-  # assets:
-  #   - images/a_dot_burr.jpeg
-  #   - images/a_dot_ham.jpeg
-  #
-  # For details regarding assets in packages, see
-  # https://flutter.dev/assets-and-images/#from-packages
-  #
-  # An image asset can refer to one or more resolution-specific "variants", see
-  # https://flutter.dev/assets-and-images/#resolution-aware.
-
-  # To add custom fonts to your plugin package, add a fonts section here,
-  # in this "flutter" section. Each entry in this list should have a
-  # "family" key with the font family name, and a "fonts" key with a
-  # list giving the asset and other descriptors for the font. For
-  # example:
-  # fonts:
-  #   - family: Schyler
-  #     fonts:
-  #       - asset: fonts/Schyler-Regular.ttf
-  #       - asset: fonts/Schyler-Italic.ttf
-  #         style: italic
-  #   - family: Trajan Pro
-  #     fonts:
-  #       - asset: fonts/TrajanPro.ttf
-  #       - asset: fonts/TrajanPro_Bold.ttf
-  #         weight: 700
-  #
-  # For details regarding fonts in packages, see
-  # https://flutter.dev/custom-fonts/#from-packages
