@@ -1,3 +1,5 @@
+import 'dart:ui' show Color;
+
 import 'package:json_annotation/json_annotation.dart';
 
 import 'enums.dart';
@@ -8,10 +10,10 @@ part 'classes.g.dart';
 @JsonSerializable(explicitToJson: true)
 class UserInfo {
   /// The user ID.
-  int uid;
+  int? uid;
 
   /// The user account.
-  String userAccount;
+  String? userAccount;
 
   /// Constructs a [UserInfo]
   UserInfo();
@@ -28,10 +30,10 @@ class UserInfo {
 @JsonSerializable(explicitToJson: true)
 class VideoDimensions {
   /// The video resolution on the horizontal axis.
-  final int width;
+  int? width;
 
   /// The video resolution on the vertical axis.
-  final int height;
+  int? height;
 
   /// Constructs a [VideoDimensions]
   VideoDimensions(this.width, this.height);
@@ -70,15 +72,15 @@ class VideoEncoderConfiguration {
   /// - The value of the dimension does not indicate the orientation mode of the output ratio. For how to set the video orientation, see [VideoOutputOrientationMode].</li>
   /// - Whether 720p+ can be supported depends on the device. If the device cannot support 720p, the frame rate will be lower than the one listed in the table.</li>
   @JsonKey(includeIfNull: false)
-  VideoDimensions dimensions;
+  VideoDimensions? dimensions;
 
   /// The video frame rate (fps). The default value is 15. You can either set the frame rate manually or choose from [VideoFrameRate]. We do not recommend setting this to a value greater than 30.
   @JsonKey(includeIfNull: false)
-  VideoFrameRate frameRate;
+  VideoFrameRate? frameRate;
 
   /// The minimum video encoder frame rate (fps). The default value is [VideoFrameRate.Min] (the SDK uses the lowest encoder frame rate).
   @JsonKey(includeIfNull: false)
-  VideoFrameRate minFrameRate;
+  VideoFrameRate? minFrameRate;
 
   /// Bitrate of the video (Kbps). Refer to the table below and set your bitrate. If you set a bitrate beyond the proper range, the SDK automatically adjusts it to a value within the range.
   ///
@@ -124,27 +126,27 @@ class VideoEncoderConfiguration {
   /// - The base bitrate in this table applies to the Communication profile.
   /// - The LiveBroadcasting profile generally requires a higher bitrate for better video quality. We recommend setting the bitrate mode as `0`. You can also set the bitrate as the base bitrate value x 2.
   @JsonKey(includeIfNull: false)
-  int bitrate;
+  int? bitrate;
 
   /// The minimum encoding bitrate (Kbps). The Agora SDK automatically adjusts the encoding bitrate to adapt to the network conditions. Using a value greater than the default value forces the video encoder to output high-quality images but may cause more packet loss and hence sacrifice the smoothness of the video transmission. That said, unless you have special requirements for image quality, Agora does not recommend changing this value.
   @JsonKey(includeIfNull: false)
-  int minBitrate;
+  int? minBitrate;
 
   /// The orientation mode.
   /// See [VideoOutputOrientationMode].
   @JsonKey(includeIfNull: false)
-  VideoOutputOrientationMode orientationMode;
+  VideoOutputOrientationMode? orientationMode;
 
   /// The video encoding degradation preference under limited bandwidth:
   /// See [DegradationPreference].
   @JsonKey(includeIfNull: false)
-  DegradationPreference degradationPrefer;
+  DegradationPreference? degradationPrefer;
 
   /// Sets the mirror mode of the published local video stream.
   /// This member only affects the video that the remote user sees.
   /// See [VideoMirrorMode].
   @JsonKey(includeIfNull: false)
-  VideoMirrorMode mirrorMode;
+  VideoMirrorMode? mirrorMode;
 
   /// Constructs a [VideoEncoderConfiguration]
   VideoEncoderConfiguration(
@@ -170,19 +172,19 @@ class VideoEncoderConfiguration {
 class BeautyOptions {
   /// The lightening contrast level, used with [lighteningLevel].
   @JsonKey(includeIfNull: false)
-  LighteningContrastLevel lighteningContrastLevel;
+  LighteningContrastLevel? lighteningContrastLevel;
 
   /// The brightness level. The value ranges between 0.0 (original) and 1.0. The default value is 0.7.
   @JsonKey(includeIfNull: false)
-  double lighteningLevel;
+  double? lighteningLevel;
 
   /// The sharpness level. The value ranges between 0.0 (original) and 1.0. The default value is 0.5. This parameter is usually used to remove blemishes.
   @JsonKey(includeIfNull: false)
-  double smoothnessLevel;
+  double? smoothnessLevel;
 
   /// The redness level. The value ranges between 0.0 (original) and 1.0. The default value is 0.1. This parameter adjusts the red saturation level.
   @JsonKey(includeIfNull: false)
-  double rednessLevel;
+  double? rednessLevel;
 
   /// Constructs a [BeautyOptions]
   BeautyOptions(
@@ -203,19 +205,19 @@ class BeautyOptions {
 @JsonSerializable(explicitToJson: true)
 class AgoraImage {
   /// HTTP/HTTPS URL address of the image on the broadcasting video. The maximum length of this parameter is 1024 bytes.
-  final String url;
+  String? url;
 
   /// Position of the image on the upper left of the broadcasting video on the horizontal axis.
-  final int x;
+  int? x;
 
   /// Position of the image on the upper left of the broadcasting video on the vertical axis.
-  final int y;
+  int? y;
 
   /// Width of the image on the broadcasting video.
-  final int width;
+  int? width;
 
   /// Height of the image on the broadcasting video.
-  final int height;
+  int? height;
 
   /// Constructs a [AgoraImage]
   AgoraImage(this.url, this.x, this.y, this.width, this.height);
@@ -232,36 +234,40 @@ class AgoraImage {
 @JsonSerializable(explicitToJson: true)
 class TranscodingUser {
   /// ID of the user in the CDN live streaming.
-  final int uid;
+  int? uid;
 
   /// Horizontal position of the video frame of the user from the top left corner of the CDN live streaming.
-  final int x;
+  int? x;
 
   /// Vertical position of the video frame of the user from the top left corner of the CDN live streaming.
-  final int y;
+  int? y;
 
   /// Width of the video frame of the user on the CDN live streaming. The default value is 360.
-  int width;
+  @JsonKey(includeIfNull: false)
+  int? width;
 
   /// Height of the video frame of the user on the CDN live streaming. The default value is 640.
-  int height;
+  @JsonKey(includeIfNull: false)
+  int? height;
 
   /// The layer index of the video frame. An integer. The value range is [0,100].
   /// - 0: (Default) Bottom layer.
   /// - 100: Top layer.
   ///
   /// **Note**: If the value is set lower than 0 or higher than 100, the [ErrorCode.InvalidArgument] error is reported.
-  int zOrder;
+  @JsonKey(includeIfNull: false)
+  int? zOrder;
 
   /// The transparency of the video frame of the user in the CDN live stream that ranges between 0.0 and 1.0. 0.0 means that the video frame is completely transparent and 1.0 means opaque. The default value is 1.0.
   @JsonKey(includeIfNull: false)
-  double alpha;
+  double? alpha;
 
   /// The audio channel ranging between 0 and 5. The default value is 0.
   /// See [AudioChannel].
   ///
   /// **Note** Special players are needed if `audioChannel` is not set as 0.
-  AudioChannel audioChannel;
+  @JsonKey(includeIfNull: false)
+  AudioChannel? audioChannel;
 
   /// Constructs a [TranscodingUser]
   TranscodingUser(
@@ -283,28 +289,6 @@ class TranscodingUser {
   Map<String, dynamic> toJson() => _$TranscodingUserToJson(this);
 }
 
-/// The background color in RGB hex. Value only. Do not include a preceding #. For example, 0xFFB6C1 (light pink). The default value is 0x000000 (black).
-@JsonSerializable(explicitToJson: true)
-class Color {
-  /// Red.
-  final int red;
-
-  /// Green.
-  final int green;
-
-  /// Blue.
-  final int blue;
-
-  /// Constructs a [Color]
-  Color(this.red, this.green, this.blue);
-
-  /// @nodoc
-  factory Color.fromJson(Map<String, dynamic> json) => _$ColorFromJson(json);
-
-  /// @nodoc
-  Map<String, dynamic> toJson() => _$ColorToJson(this);
-}
-
 /// A class for managing user-specific CDN live audio/video transcoding settings.
 @JsonSerializable(explicitToJson: true)
 class LiveTranscoding {
@@ -312,75 +296,75 @@ class LiveTranscoding {
   /// - When pushing video streams to the CDN, ensure that `width` is at least 64; otherwise, the Agora server adjusts the value to 64.
   /// - When pushing audio streams to the CDN, set `width` and `height` as 0.
   @JsonKey(includeIfNull: false)
-  int width;
+  int? width;
 
   /// Height (pixel) of the video. The default value is 640.
   /// - When pushing video streams to the CDN, ensure that `height` is at least 64; otherwise, the Agora server adjusts the value to 64.
   /// - When pushing audio streams to the CDN, set `width` and `height` as 0.
   @JsonKey(includeIfNull: false)
-  int height;
+  int? height;
 
   /// Bitrate (Kbps) of the CDN live output video stream. The default value is 400. Set this parameter according to the [VideoEncoderConfiguration.bitrate](Video Bitrate Table). If you set a bitrate beyond the proper range, the SDK automatically adapts it to a value within the range.
   @JsonKey(includeIfNull: false)
-  int videoBitrate;
+  int? videoBitrate;
 
   /// The frame rate (fps) of the video. The value range is (0, 30]. The default value is 15. The Agora server adjusts any value over 30 to 30.
   @JsonKey(includeIfNull: false)
-  VideoFrameRate videoFramerate;
+  VideoFrameRate? videoFramerate;
 
   /// Agora does not recommend using this parameter.
   /// - `true`: Low latency with unassured quality.
   /// - `false`: (Default) High latency with assured quality.
   @deprecated
   @JsonKey(includeIfNull: false)
-  bool lowLatency;
+  bool? lowLatency;
 
   /// Gop of the video frames in the CDN live stream. The default value is 30 fps.
   @JsonKey(includeIfNull: false)
-  int videoGop;
+  int? videoGop;
 
   /// The watermark image added to the CDN live publishing stream. Ensure that the format of the image is PNG. Once a watermark image is added, the audience of the CDN live publishing stream can see it.
   /// See [AgoraImage].
   @JsonKey(includeIfNull: false)
-  AgoraImage watermark;
+  AgoraImage? watermark;
 
   /// The background image added to the CDN live publishing stream. Once a background image is added, the audience of the CDN live publishing stream can see it.
   /// See [AgoraImage].
   @JsonKey(includeIfNull: false)
-  AgoraImage backgroundImage;
+  AgoraImage? backgroundImage;
 
   /// Self-defined audio-sample rate: [AudioSampleRateType].
   @JsonKey(includeIfNull: false)
-  AudioSampleRateType audioSampleRate;
+  AudioSampleRateType? audioSampleRate;
 
   /// Bitrate (Kbps) of the CDN live audio output stream. The default value is 48 and the highest value is 128.
   @JsonKey(includeIfNull: false)
-  int audioBitrate;
+  int? audioBitrate;
 
   /// Agora’s self-defined audio channel type. Agora recommends choosing 1 (mono), or 2 (stereo) audio channels. Special players are required if you choose 3, 4, or 5.
   /// See [AudioChannel].
   @JsonKey(includeIfNull: false)
-  AudioChannel audioChannels;
+  AudioChannel? audioChannels;
 
   /// Audio codec profile type: [AudioCodecProfileType]. Set it as `LCAAC` or `HEAAC`. The default value is `LCAAC`.
   @JsonKey(includeIfNull: false)
-  AudioCodecProfileType audioCodecProfile;
+  AudioCodecProfileType? audioCodecProfile;
 
   /// Video codec profile type: [VideoCodecProfileType]. Set it as `BASELINE`, `MAIN`, or `HIGH` (default). If you set this parameter to other values, Agora adjusts it to the default value `HIGH`.
   @JsonKey(includeIfNull: false)
-  VideoCodecProfileType videoCodecProfile;
+  VideoCodecProfileType? videoCodecProfile;
 
   /// The background color in RGB hex. Value only. Do not include a preceding #. For example, 0xFFB6C1 (light pink). The default value is 0x000000 (black).
-  /// See [Color].
-  @JsonKey(includeIfNull: false)
-  Color backgroundColor;
+  @JsonKey(
+      includeIfNull: false, fromJson: _$ColorFromJson, toJson: _$ColorToJson)
+  Color? backgroundColor;
 
   /// Reserved property. Extra user-defined information to send the Supplemental Enhancement Information (SEI) for the H.264/H.265 video stream to the CDN live client. Maximum length: 4096 Bytes.
   @JsonKey(includeIfNull: false)
-  String userConfigExtraInfo;
+  String? userConfigExtraInfo;
 
   /// An TranscodingUser object managing the user layout configuration in the CDN live stream. Agora supports a maximum of 17 transcoding users in a CDN live stream channel.
-  final List<TranscodingUser> transcodingUsers;
+  List<TranscodingUser>? transcodingUsers;
 
   /// Constructs a [LiveTranscoding]
   LiveTranscoding(
@@ -408,6 +392,18 @@ class LiveTranscoding {
 
   /// @nodoc
   Map<String, dynamic> toJson() => _$LiveTranscodingToJson(this);
+
+  static Color _$ColorFromJson(Map<String, dynamic> json) => Color.fromRGBO(
+      json['red'] as int, json['green'] as int, json['blue'] as int, 1.0);
+
+  static Map<String, dynamic>? _$ColorToJson(Color? instance) =>
+      instance != null
+          ? <String, dynamic>{
+              'red': instance.red,
+              'green': instance.green,
+              'blue': instance.blue,
+            }
+          : null;
 }
 
 /// The ChannelMediaInfo class.
@@ -415,14 +411,14 @@ class LiveTranscoding {
 class ChannelMediaInfo {
   /// The channel name.
   @JsonKey(includeIfNull: false)
-  String channelName;
+  String? channelName;
 
   /// The token that enables the user to join the channel.
   @JsonKey(includeIfNull: false)
-  String token;
+  String? token;
 
   /// The user ID.
-  final int uid;
+  int? uid;
 
   /// Constructs a [ChannelMediaInfo]
   ChannelMediaInfo(this.uid, {this.channelName, this.token});
@@ -444,7 +440,7 @@ class ChannelMediaRelayConfiguration {
   /// - `token`: The token for joining the source channel. It is generated with the `channelName` and `uid` you set in `srcInfo`.
   ///   - If you have not enabled the App Certificate, set this parameter as the default value NULL, which means the SDK applies the App ID.
   ///   - If you have enabled the App Certificate, you must use the token generated with the `channelName` and `uid`.
-  final ChannelMediaInfo srcInfo;
+  ChannelMediaInfo? srcInfo;
 
   /// The information of the destination channel: [ChannelMediaInfo]. It contains the following members:
   ///- `channelName`: The name of the destination channel.
@@ -453,7 +449,7 @@ class ChannelMediaRelayConfiguration {
   ///  - `token`: The token for joining the destination channel. It is generated with the `channelName` and `uid` you set in `destInfo`.
   ///    - If you have not enabled the App Certificate, set this parameter as the default value NULL, which means the SDK applies the App ID.
   ///    - If you have enabled the App Certificate, you must use the token generated with the `channelName` and `uid`.
-  final List<ChannelMediaInfo> destInfos;
+  List<ChannelMediaInfo>? destInfos;
 
   /// Constructs a [ChannelMediaRelayConfiguration]
   ChannelMediaRelayConfiguration(this.srcInfo, this.destInfos);
@@ -470,16 +466,16 @@ class ChannelMediaRelayConfiguration {
 @JsonSerializable(explicitToJson: true)
 class LastmileProbeConfig {
   /// Whether to probe uplink of lastmile. i.e., audience don't need probe uplink bandwidth.
-  final bool probeUplink;
+  bool? probeUplink;
 
   /// Whether to probe downlink of lastmile.
-  final bool probeDownlink;
+  bool? probeDownlink;
 
   /// The expected maximum sending bitrate in bps in range of [100000, 5000000]. It is recommended to set this value according to the required bitrate of selected video profile.
-  final int expectedUplinkBitrate;
+  int? expectedUplinkBitrate;
 
   /// The expected maximum receive bitrate in bps in range of [100000, 5000000].
-  final int expectedDownlinkBitrate;
+  int? expectedDownlinkBitrate;
 
   /// Constructs a [LastmileProbeConfig]
   LastmileProbeConfig(this.probeUplink, this.probeDownlink,
@@ -497,16 +493,16 @@ class LastmileProbeConfig {
 @JsonSerializable(explicitToJson: true)
 class Rectangle {
   /// The horizontal offset from the top-left corner.
-  final int x;
+  int? x;
 
   /// The vertical offset from the top-left corner.
-  final int y;
+  int? y;
 
   /// The width (pixels) of the watermark image.
-  final int width;
+  int? width;
 
   /// The height (pixels) of the watermark image.
-  final int height;
+  int? height;
 
   /// Constructs a [Rectangle]
   Rectangle(this.x, this.y, this.width, this.height);
@@ -526,15 +522,15 @@ class WatermarkOptions {
   /// - `true`: (Default) The watermark image is visible in preview.
   /// - `false`: The watermark image is not visible in preview.
   @JsonKey(includeIfNull: false)
-  bool visibleInPreview;
+  bool? visibleInPreview;
 
   /// The watermark position in the landscape mode.
   /// See [Rectangle].
-  final Rectangle positionInLandscapeMode;
+  Rectangle? positionInLandscapeMode;
 
   /// The watermark position in the portrait mode.
   /// See [Rectangle].
-  final Rectangle positionInPortraitMode;
+  Rectangle? positionInPortraitMode;
 
   /// Constructs a [WatermarkOptions]
   WatermarkOptions(this.positionInLandscapeMode, this.positionInPortraitMode,
@@ -552,40 +548,42 @@ class WatermarkOptions {
 @JsonSerializable(explicitToJson: true)
 class LiveInjectStreamConfig {
   /// Width of the added stream to the broadcast. The default value is 0, which is the same width as the original stream.
-  int width;
+  @JsonKey(includeIfNull: false)
+  int? width;
 
   /// Height of the added stream to the broadcast. The default value is 0, which is the same height as the original stream.
-  int height;
+  @JsonKey(includeIfNull: false)
+  int? height;
 
   /// Video GOP of the added stream to the broadcast. The default value is 30 frames.
   @JsonKey(includeIfNull: false)
-  int videoGop;
+  int? videoGop;
 
   /// Video frame rate of the added stream to the broadcast. The default value is 15 fps.
   @JsonKey(includeIfNull: false)
-  VideoFrameRate videoFramerate;
+  VideoFrameRate? videoFramerate;
 
   /// Video bitrate of the added stream to the broadcast. The default value is 400 Kbps.
   @JsonKey(includeIfNull: false)
-  int videoBitrate;
+  int? videoBitrate;
 
   /// Audio sample rate of the added stream to the broadcast: [AudioSampleRateType]. The default value is 44100 Hz.
   ///
   /// **Note** We recommend you use the default value and not reset it.
   @JsonKey(includeIfNull: false)
-  AudioSampleRateType audioSampleRate;
+  AudioSampleRateType? audioSampleRate;
 
   /// Audio bitrate of the added stream to the broadcast. The default value is 48.
   ///
   /// **Note** We recommend you use the default value and not reset it.
   @JsonKey(includeIfNull: false)
-  int audioBitrate;
+  int? audioBitrate;
 
   /// Audio channels to add into the broadcast. The value ranges between 1 and 2. The default value is 1.
   ///
   /// **Note** We recommend you use the default value and not reset it.
   @JsonKey(includeIfNull: false)
-  AudioChannel audioChannels;
+  AudioChannel? audioChannels;
 
   /// Constructs a [LiveInjectStreamConfig]
   LiveInjectStreamConfig(
@@ -606,19 +604,30 @@ class LiveInjectStreamConfig {
   Map<String, dynamic> toJson() => _$LiveInjectStreamConfigToJson(this);
 }
 
-/// The definition of CameraCapturerConfiguration.
+/// The configuration of camera capturer.
 @JsonSerializable(explicitToJson: true)
 class CameraCapturerConfiguration {
-  /// The camera capturer configuration.
+  /// The camera capture configuration.
   /// See [CameraCaptureOutputPreference].
-  final CameraCaptureOutputPreference preference;
+  CameraCaptureOutputPreference? preference;
+
+  /// The width (px) of the video image captured by the local camera.
+  /// To customize the width of the video image, set [preference] as `Manual(3)` first, and then use this parameter.
+  @JsonKey(includeIfNull: false)
+  int? captureWidth;
+
+  /// The height (px) of the video image captured by the local camera.
+  /// To customize the height of the video image, set [preference] as `Manual(3)` first, and then use this parameter.
+  @JsonKey(includeIfNull: false)
+  int? captureHeight;
 
   /// The camera direction.
   /// See [CameraDirection].
-  final CameraDirection cameraDirection;
+  CameraDirection? cameraDirection;
 
   /// Constructs a [CameraCapturerConfiguration]
-  CameraCapturerConfiguration(this.preference, this.cameraDirection);
+  CameraCapturerConfiguration(this.preference, this.cameraDirection,
+      {this.captureWidth, this.captureHeight});
 
   /// @nodoc
   factory CameraCapturerConfiguration.fromJson(Map<String, dynamic> json) =>
@@ -637,7 +646,7 @@ class ChannelMediaOptions {
   ///
   /// This member serves a similar function to the [RtcEngine.muteAllRemoteAudioStreams] method.
   /// After joining the channel, you can call `muteAllRemoteAudioStreams` to set whether to subscribe to audio streams in the channel.
-  final bool autoSubscribeAudio;
+  bool? autoSubscribeAudio;
 
   /// Determines whether to subscribe to video streams when the user joins the channel.
   /// - `true`: (Default) Subscribe.
@@ -645,7 +654,7 @@ class ChannelMediaOptions {
   ///
   /// This member serves a similar function to the [RtcEngine.muteAllRemoteVideoStreams] method.
   /// After joining the channel, you can call `muteAllRemoteVideoStreams` to set whether to subscribe to video streams in the channel.
-  final bool autoSubscribeVideo;
+  bool? autoSubscribeVideo;
 
   /// Constructs a [ChannelMediaOptions]
   ChannelMediaOptions(this.autoSubscribeAudio, this.autoSubscribeVideo);
@@ -659,19 +668,17 @@ class ChannelMediaOptions {
 }
 
 /// Definition of `EncryptionConfig`.
-///
-///
 @JsonSerializable(explicitToJson: true)
 class EncryptionConfig {
   /// Encryption mode. The default encryption mode is `AES128XTS`. See [EncryptionMode].
-  final EncryptionMode encryptionMode;
+  EncryptionMode? encryptionMode;
 
   /// Encryption key in string type.
   ///
   /// **Note**
   ///
   /// If you do not set an encryption key or set it as null, you cannot use the built-in encryption, and the SDK returns [ErrorCode.InvalidArgument].
-  final String encryptionKey;
+  String? encryptionKey;
 
   /// Constructs a [EncryptionConfig]
   EncryptionConfig(this.encryptionMode, this.encryptionKey);
@@ -688,83 +695,91 @@ class EncryptionConfig {
 @JsonSerializable(explicitToJson: true)
 class RtcStats {
   /// Call duration in seconds, represented by an aggregate value.
-  int totalDuration;
+  int? totalDuration;
 
   /// Total number of bytes transmitted, represented by an aggregate value.
-  int txBytes;
+  int? txBytes;
 
   /// Total number of bytes received, represented by an aggregate value.
-  int rxBytes;
+  int? rxBytes;
 
   /// Total number of audio bytes sent (bytes), represented by an aggregate value.
-  int txAudioBytes;
+  int? txAudioBytes;
 
   /// Total number of video bytes sent (bytes), represented by an aggregate value.
-  int txVideoBytes;
+  int? txVideoBytes;
 
   /// Total number of audio bytes received (bytes), represented by an aggregate value.
-  int rxAudioBytes;
+  int? rxAudioBytes;
 
   /// Total number of video bytes received (bytes), represented by an aggregate value.
-  int rxVideoBytes;
+  int? rxVideoBytes;
 
   /// Transmission bitrate in Kbps, represented by an instantaneous value.
-  int txKBitRate;
+  int? txKBitRate;
 
   /// Receive bitrate (Kbps), represented by an instantaneous value.
-  int rxKBitRate;
+  int? rxKBitRate;
 
   /// The transmission bitrate of the audio packet (Kbps), represented by an instantaneous value.
-  int txAudioKBitRate;
+  int? txAudioKBitRate;
 
   /// Audio receive bitrate (Kbps), represented by an instantaneous value.
-  int rxAudioKBitRate;
+  int? rxAudioKBitRate;
 
   /// Video transmission bitrate (Kbps), represented by an instantaneous value.
-  int txVideoKBitRate;
+  int? txVideoKBitRate;
 
   /// Video receive bitrate (Kbps), represented by an instantaneous value.
-  int rxVideoKBitRate;
+  int? rxVideoKBitRate;
 
   /// The number of users in the channel.
   /// - Communication profile: The number of users in the channel.
   /// - Live Broadcast profile:
   ///   - If the local user is an audience: The number of users in the channel = The number of hosts in the channel + 1.
   ///   - If the local user is a host: The number of users in the channel = The number of hosts in the channel.
-  int users;
+  int? users;
 
   /// Client-server latency.
-  int lastmileDelay;
+  int? lastmileDelay;
 
   /// The packet loss rate (%) from the local client to Agora's edge server, before network countermeasures.
-  int txPacketLossRate;
+  int? txPacketLossRate;
 
   /// The packet loss rate (%) from Agora's edge server to the local client, before network countermeasures.
-  int rxPacketLossRate;
+  int? rxPacketLossRate;
 
   /// System CPU usage (%).
-  double cpuTotalUsage;
+  ///
+  /// **Note**
+  ///
+  /// The `cpuTotalUsage` reported in the `leaveChannel` callback is always 0.
+  double? cpuTotalUsage;
 
   /// Application CPU usage (%).
-  double cpuAppUsage;
+  ///
+  /// **Note**
+  ///
+  /// The `cpuAppUsage` reported in the `leaveChannel` callback is always 0.
+  double? cpuAppUsage;
 
   /// The round-trip time delay from the client to the local router.
-  int gatewayRtt;
+  int? gatewayRtt;
 
   /// The memory usage ratio of the app (%).
   ///
   /// **Note**: This value is for reference only. Due to system limitations, you may not get the value of this member.
-  double memoryAppUsageRatio;
+  double? memoryAppUsageRatio;
 
   /// The memory usage ratio of the system (%).
   ///
   /// **Note**: This value is for reference only. Due to system limitations, you may not get the value of this member.
-  double memoryTotalUsageRatio;
+  double? memoryTotalUsageRatio;
 
   /// The memory usage of the app (KB).
   ///
   /// **Note**: This value is for reference only. Due to system limitations, you may not get the value of this member.
-  int memoryAppUsageInKbytes;
+  int? memoryAppUsageInKbytes;
 
   /// Constructs a [RtcStats]
   RtcStats();
@@ -782,24 +797,23 @@ class RtcStats {
 @JsonSerializable(explicitToJson: true)
 class AudioVolumeInfo {
   /// The user ID of the speaker. The uid of the local user is 0.
-  int uid;
+  int? uid;
 
   /// The sum of the voice volume and audio-mixing volume of the speaker. The value ranges between 0 (lowest volume) and 255 (highest volume).
-  int volume;
+  int? volume;
 
   /// Voice activity status of the local user.
   /// - 0: The local user is not speaking.
   /// - 1: The local user is speaking.
   ///
-  ///
   /// **Note**
   /// - The `vad` parameter cannot report the voice activity status of the remote users. In the remote users' callback, `vad` = 0.
   /// - Ensure that you set `report_vad`(true) in the [RtcEngine.enableAudioVolumeIndication] method to enable the voice activity
   /// detection of the local user.
-  int vad;
+  int? vad;
 
   /// The channel ID, which indicates which channel the speaker is in.
-  String channelId;
+  String? channelId;
 
   /// Constructs a [AudioVolumeInfo]
   AudioVolumeInfo();
@@ -816,16 +830,16 @@ class AudioVolumeInfo {
 @JsonSerializable(explicitToJson: true)
 class Rect {
   /// The X coordinate of the left side of the rectangle.
-  int left;
+  int? left;
 
   /// The Y coordinate of the top side of the rectangle.
-  int top;
+  int? top;
 
   /// The X coordinate of the right side of the rectangle.
-  int right;
+  int? right;
 
   /// The Y coordinate of the bottom side of the rectangle.
-  int bottom;
+  int? bottom;
 
   /// Constructs a [Rect]
   Rect();
@@ -841,13 +855,13 @@ class Rect {
 @JsonSerializable(explicitToJson: true)
 class LastmileProbeOneWayResult {
   /// The packet loss rate (%).
-  int packetLossRate;
+  int? packetLossRate;
 
   /// The network jitter (ms).
-  int jitter;
+  int? jitter;
 
   /// The estimated available bandwidth (bps).
-  int availableBandwidth;
+  int? availableBandwidth;
 
   /// Constructs a [LastmileProbeOneWayResult]
   LastmileProbeOneWayResult();
@@ -865,18 +879,18 @@ class LastmileProbeOneWayResult {
 class LastmileProbeResult {
   /// The state of the probe test.
   /// See [LastmileProbeResultState].
-  LastmileProbeResultState state;
+  LastmileProbeResultState? state;
 
   /// The round-trip delay time (ms).
-  int rtt;
+  int? rtt;
 
   /// The uplink last-mile network report.
   /// See [LastmileProbeOneWayResult].
-  LastmileProbeOneWayResult uplinkReport;
+  LastmileProbeOneWayResult? uplinkReport;
 
   /// The downlink last-mile network report.
   /// See [LastmileProbeOneWayResult].
-  LastmileProbeOneWayResult downlinkReport;
+  LastmileProbeOneWayResult? downlinkReport;
 
   /// Constructs a [LastmileProbeResult]
   LastmileProbeResult();
@@ -893,18 +907,16 @@ class LastmileProbeResult {
 @JsonSerializable(explicitToJson: true)
 class LocalAudioStats {
   /// The number of channels.
-  int numChannels;
+  int? numChannels;
 
   /// The sample rate (Hz).
-  int sentSampleRate;
+  int? sentSampleRate;
 
   /// The average sending bitrate (Kbps).
-  int sentBitrate;
+  int? sentBitrate;
 
   /// The video packet loss rate (%) from the local client to the Agora edge server before applying the anti-packet loss strategies.
-  ///
-  ///
-  int txPacketLossRate;
+  int? txPacketLossRate;
 
   /// Constructs a [LocalAudioStats]
   LocalAudioStats();
@@ -921,52 +933,51 @@ class LocalAudioStats {
 @JsonSerializable(explicitToJson: true)
 class LocalVideoStats {
   /// Bitrate (Kbps) sent in the reported interval, which does not include the bitrate of the re-transmission video after the packet loss.
-  int sentBitrate;
+  int? sentBitrate;
 
   /// Frame rate (fps) sent in the reported interval, which does not include the frame rate of the re-transmission video after the packet loss.
-  int sentFrameRate;
+  int? sentFrameRate;
 
   /// The encoder output frame rate (fps) of the local video.
-  int encoderOutputFrameRate;
+  int? encoderOutputFrameRate;
 
   /// The renderer output frame rate (fps) of the local video.
-  int rendererOutputFrameRate;
+  int? rendererOutputFrameRate;
 
   /// The target bitrate (Kbps) of the current encoder. This value is estimated by the SDK based on the current network conditions.
-  int targetBitrate;
+  int? targetBitrate;
 
   /// The target frame rate (fps) of the current encoder.
-  int targetFrameRate;
+  int? targetFrameRate;
 
   /// Quality change of the local video in terms of target frame rate and target bit rate since last count.
   /// See [VideoQualityAdaptIndication].
-  VideoQualityAdaptIndication qualityAdaptIndication;
+  VideoQualityAdaptIndication? qualityAdaptIndication;
 
   /// The encoding bitrate (Kbps), which does not include the bitrate of the re-transmission video after packet loss.
-  int encodedBitrate;
+  int? encodedBitrate;
 
   /// The width of the encoding frame (px).
-  int encodedFrameWidth;
+  int? encodedFrameWidth;
 
   /// The height of the encoding frame (px).
-  int encodedFrameHeight;
+  int? encodedFrameHeight;
 
   /// The value of the sent frame rate, represented by an aggregate value.
-  int encodedFrameCount;
+  int? encodedFrameCount;
 
   /// The codec type of the local video.
   /// See [VideoCodecType].
-  VideoCodecType codecType;
+  VideoCodecType? codecType;
 
   /// The video packet loss rate (%) from the local client to the Agora edge server before applying the anti-packet loss strategies.
-  ///
-  ///
-  int txPacketLossRate;
+  int? txPacketLossRate;
 
   /// The capture frame rate (fps) of the local video.
-  ///
-  ///
-  int captureFrameRate;
+  int? captureFrameRate;
+
+  /// The capture brightness level type.
+  CaptureBrightnessLevelType? captureBrightnessLevel;
 
   /// Constructs a [LocalVideoStats]
   LocalVideoStats();
@@ -983,47 +994,92 @@ class LocalVideoStats {
 @JsonSerializable(explicitToJson: true)
 class RemoteAudioStats {
   /// ID of the user sending the audio streams.
-  int uid;
+  int? uid;
 
   /// Audio quality received by the user.
   /// See [NetworkQuality].
-  NetworkQuality quality;
+  NetworkQuality? quality;
 
   /// Network delay (ms) from the sender to the receiver.
-  int networkTransportDelay;
+  int? networkTransportDelay;
 
   /// Network delay (ms) from the receiver to the jitter buffer.
   ///
   /// **Note**
   ///
   /// When the receiver is an audience member and [AudienceLatencyLevelType] is `1`, this parameter does not take effect.
-  int jitterBufferDelay;
+  int? jitterBufferDelay;
 
   /// Packet loss rate in the reported interval.
-  int audioLossRate;
+  int? audioLossRate;
 
   /// The number of channels.
-  int numChannels;
+  int? numChannels;
 
   /// The sample rate (Hz) of the received audio stream in the reported interval.
-  int receivedSampleRate;
+  int? receivedSampleRate;
 
   /// The average bitrate (Kbps) of the received audio stream in the reported interval.
-  int receivedBitrate;
+  int? receivedBitrate;
 
   /// The total freeze time (ms) of the remote audio stream after the remote user joins the channel. In the reported interval, audio freeze occurs when the audio frame loss rate reaches 4%. totalFrozenTime = The audio freeze time × 2 × 1000 (ms).
-  int totalFrozenTime;
+  int? totalFrozenTime;
 
   /// The total audio freeze time as a percentage (%) of the total time when the audio is available.
-  int frozenRate;
+  int? frozenRate;
 
   /// The total time (ms) when the remote user in the Communication profile or the remote broadcaster in the LiveBroadcasting profile neither stops sending the audio stream nor disables the audio module after joining the channel.
-  int totalActiveTime;
+  int? totalActiveTime;
 
   /// The total active time (ms) of the remote audio stream after the remote user publish the audio stream.
+  int? publishDuration;
+
+  /// Quality of experience (QoE) of the local user when receiving a remote audio stream. See [ExperienceQuality].
+  ExperienceQualityType? qoeQuality;
+
+  /// The reason for poor QoE of the local user when receiving a remote audio stream. See [ExperiencePoorReason].
+  ExperiencePoorReason? qualityChangedReason;
+
+  /// The [quality] of the remote audio stream as determined by the Agora real-time audio MOS (Mean Opinion Score) measurement method in the reported interval.
+  /// The return value ranges from 0 to 500. Dividing the return value by 100 gets the MOS score, which ranges from 0 to 5. The higher the score, the better the audio quality.
   ///
+  /// The subjective perception of audio quality corresponding to the Agora real-time audio MOS scores is as follows:
   ///
-  int publishDuration;
+  /// <table border="1">
+  /// <thead>
+  /// <tr>
+  ///   <th>MOS score</th>
+  ///   <th>Perception of audio quality</th>
+  /// </tr>
+  /// </thead>
+  /// <tbody>
+  /// <tr>
+  /// <td>Greater than 4</td>
+  /// <td>- Excellent. The audio sounds clear and smooth.</td>
+  /// </tr>
+  /// <tr>
+  /// <td>From 3.5 to 4</td>
+  /// <td>Good. The audio has some perceptible impairment, but still sounds clear.</td>
+  /// </tr>
+  /// <tr>
+  /// <td>From 3 to 3.5</td>
+  /// <td>Fair. The audio freezes occasionally and requires attentive listening.</td>
+  /// </tr>
+  /// <tr>
+  /// <td>From 2.5 to 3</td>
+  /// <td>Poor. The audio sounds choppy and requires considerable effort to understand.</td>
+  /// </tr>
+  /// <tr>
+  /// <td>From 2 to 2.5</td>
+  /// <td>Bad. The audio has occasional noise. Consecutive audio dropouts occur, resulting in some information loss. The users can communicate only with difficulty.</td>
+  /// </tr>
+  /// <tr>
+  /// <td>Less than 2</td>
+  /// <td> Very bad. The audio has persistent noise. Consecutive audio dropouts are frequent, resulting in severe information loss. Communication is nearly impossible.</td>
+  /// </tr>
+  /// </tbody>
+  /// </table>
+  int? mosValue;
 
   /// Constructs a [RemoteAudioStats]
   RemoteAudioStats();
@@ -1040,48 +1096,46 @@ class RemoteAudioStats {
 @JsonSerializable(explicitToJson: true)
 class RemoteVideoStats {
   /// ID of the user sending the video streams.
-  int uid;
+  int? uid;
 
   /// Time delay (ms). In scenarios where audio and video is synchronized, you can use the value of `networkTransportDelay` and `jitterBufferDelay` in [RemoteAudioStats] to know the delay statistics of the remote video.
   @deprecated
-  int delay;
+  int? delay;
 
   /// Width (pixels) of the remote video.
-  int width;
+  int? width;
 
   /// Height (pixels) of the remote video.
-  int height;
+  int? height;
 
   /// Bitrate (Kbps) received in the reported interval.
-  int receivedBitrate;
+  int? receivedBitrate;
 
   /// The decoder output frame rate (fps) of the remote video.
-  int decoderOutputFrameRate;
+  int? decoderOutputFrameRate;
 
   /// The renderer output frame rate (fps) of the remote video.
-  int rendererOutputFrameRate;
+  int? rendererOutputFrameRate;
 
   /// Packet loss rate (%) of the remote video stream after network countermeasures.
-  int packetLossRate;
+  int? packetLossRate;
 
   /// Video stream type (high-stream or low-stream).
   /// See [VideoStreamType].
-  VideoStreamType rxStreamType;
+  VideoStreamType? rxStreamType;
 
   /// The total freeze time (ms) of the remote video stream after the remote user joins the channel.
   /// In a video session where the frame rate is set to no less than 5 fps, video freeze occurs when the time interval between two adjacent renderable video frames is more than 500 ms.
-  int totalFrozenTime;
+  int? totalFrozenTime;
 
   /// The total video freeze time (`totalFrozenTime`) as a percentage (%) of the total time when the video is available (`totalActiveTime`).
-  int frozenRate;
+  int? frozenRate;
 
   /// The total time (ms) when the remote user in the Communication profile or the remote broadcaster in the Live-broadcast profile neither stops sending the video stream nor disables the video module after joining the channel.
-  int totalActiveTime;
+  int? totalActiveTime;
 
   /// The total publish duration (ms) of the remote video stream.
-  ///
-  ///
-  int publishDuration;
+  int? publishDuration;
 
   /// Constructs a [RemoteVideoStats]
   RemoteVideoStats();
@@ -1098,19 +1152,19 @@ class RemoteVideoStats {
 @JsonSerializable(explicitToJson: true)
 class FacePositionInfo {
   /// The x coordinate (px) of the human face in the local video. Taking the top left corner of the captured video as the origin, the x coordinate represents the relative lateral displacement of the top left corner of the human face to the origin.
-  int x;
+  int? x;
 
   /// The y coordinate (px) of the human face in the local video. Taking the top left corner of the captured video as the origin, the y coordinate represents the relative longitudinal displacement of the top left corner of the human face to the origin.
-  int y;
+  int? y;
 
   /// The width (px) of the human face in the captured video.
-  int width;
+  int? width;
 
   /// The height (px) of the human face in the captured video.
-  int height;
+  int? height;
 
   /// The distance (cm) between the human face and the screen.
-  int distance;
+  int? distance;
 
   /// Constructs a [FacePositionInfo]
   FacePositionInfo();
@@ -1127,10 +1181,10 @@ class FacePositionInfo {
 @JsonSerializable(explicitToJson: true)
 class ClientRoleOptions {
   /// The latency level of an audience member in a live interactive streaming.
-  AudienceLatencyLevelType audienceLatencyLevel;
+  AudienceLatencyLevelType? audienceLatencyLevel;
 
   /// Constructs a [ClientRoleOptions]
-  ClientRoleOptions();
+  ClientRoleOptions(this.audienceLatencyLevel);
 
   /// @nodoc
   factory ClientRoleOptions.fromJson(Map<String, dynamic> json) =>
@@ -1138,4 +1192,112 @@ class ClientRoleOptions {
 
   /// @nodoc
   Map<String, dynamic> toJson() => _$ClientRoleOptionsToJson(this);
+}
+
+///  Log file configurations.
+@JsonSerializable(explicitToJson: true)
+class LogConfig {
+  /// The absolute path of log files. The default file path is `/storage/emulated/0/Android/data/<package name>/files/agorasdk.log` for Android or `App Sandbox/Library/caches/agorasdk.log` for iOS.
+  ///
+  /// Ensure that the directory for the log files exists and is writable. You can use this parameter to rename the log files.
+  @JsonKey(includeIfNull: false)
+  String? filePath;
+
+  /// The size (KB) of a log file. The default value is 1024 KB.
+  ///
+  /// If you set `fileSize` to 1024 KB, the SDK outputs at most 5 MB log files; if you set it to less than 1024 KB, the setting is invalid, and the maximum size of a log file is still 1024 KB.
+  @JsonKey(includeIfNull: false)
+  int? fileSize;
+
+  /// The output log level of the SDK. See details in [LogLevel].
+  ///
+  /// For example, if you set the log level to `Warn`, the SDK outputs the logs within levels `Fatal`, `Error`, and `Warn`.
+  @JsonKey(includeIfNull: false)
+  LogLevel? level;
+
+  /// Constructs a [LogConfig]
+  LogConfig({this.filePath, this.fileSize, this.level});
+
+  /// @nodoc
+  factory LogConfig.fromJson(Map<String, dynamic> json) =>
+      _$LogConfigFromJson(json);
+
+  /// @nodoc
+  Map<String, dynamic> toJson() => _$LogConfigToJson(this);
+}
+
+///  The configurations for the data stream.
+///
+/// The following table shows the relationship between the [syncWithAudio] parameter and the [ordered] parameter:
+///
+/// | [syncWithAudio] | [ordered] | SDK behaviors                                                                                                                                                                                                                                                                                                                                                                                |
+/// |---------------|---------|----------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------|
+/// | `false`       | `false`    | The SDK triggers the `streamMessage` callback immediately after the receiver receives a data packet.                                                                                                                                                                                                                                                                             |
+/// | `true`        | `false`    | If the data packet delay is within the audio delay, the SDK triggers the `streamMessage` callback when the synchronized audio packet is played out. If the data packet delay exceeds the audio delay, the SDK triggers the `streamMessage` callback as soon as the data packet is received. In this case, the data packet is not synchronized with the audio packet. |
+/// | `false`       | `true`     | If the delay of a data packet is within five seconds, the SDK corrects the order of the data packet. If the delay of a data packet exceeds five seconds, the SDK discards the data packet.                                                                                                                                                                                                   |
+/// | `true`        | `true`     | If the delay of a data packet is within the audio delay, the SDK corrects the order of the data packet. If the delay of a data packet exceeds the audio delay, the SDK discards this data packet.                                                                                                                                                                                            |
+@JsonSerializable(explicitToJson: true)
+class DataStreamConfig {
+  /// Whether to synchronize the data packet with the published audio packet.
+  ///
+  /// - `true`: Synchronize the data packet with the audio packet.
+  /// - `false`: Do not synchronize the data packet with the audio packet.
+  ///
+  /// When you set the data packet to synchronize with the audio, then if the data packet delay is within the audio delay range, the SDK triggers the `streamMessage` callback when the synchronized audio packet is played out.
+  /// Do not set this parameter as `true` if you need the receiver to receive the data packet immediately. Agora recommends that you set this parameter to `true` only when you need to implement specific functions, for example lyric synchronization.
+  @JsonKey(includeIfNull: false)
+  bool? syncWithAudio;
+
+  /// Whether the SDK guarantees that the receiver receives the data in the sent order.
+  ///
+  /// - `true`: Guarantee that the receiver receives the data in the sent order.
+  /// - `false`: Do not guarantee that the receiver receives the data in the sent order.
+  ///
+  /// Do not set this parameter to `true` if you need the receiver to receive the data immediately.
+  @JsonKey(includeIfNull: false)
+  bool? ordered;
+
+  /// Constructs a [DataStreamConfig]
+  DataStreamConfig({this.syncWithAudio, this.ordered});
+
+  /// @nodoc
+  factory DataStreamConfig.fromJson(Map<String, dynamic> json) =>
+      _$DataStreamConfigFromJson(json);
+
+  /// @nodoc
+  Map<String, dynamic> toJson() => _$DataStreamConfigToJson(this);
+}
+
+///  Configurations for the RtcEngineConfig instance.
+@JsonSerializable(explicitToJson: true)
+class RtcEngineConfig {
+  /// The App ID issued to you by Agora. See [How to get the App ID](https://docs.agora.io/en/Agora%20Platform/token#get-an-app-id).
+  /// Only users in apps with the same App ID can join the same channel and communicate with each other. Use an App ID to create only
+  /// one `RtcEngine` instance. To change your App ID, call `destroy` to destroy the current `RtcEngine` instance and then call `createWithConfig`
+  /// to create an `RtcEngine` instance with the new App ID.
+  String? appId;
+
+  /// The region for connection. This advanced feature applies to scenarios that have regional restrictions.
+  ///
+  /// For the regions that Agora supports, see [AreaCode]. After specifying the region, the SDK connects to the Agora servers within that region.
+  @JsonKey(includeIfNull: false)
+  AreaCode? areaCode;
+
+  /// The configuration of the log files that the SDK outputs. See [LogConfig].
+  ///
+  /// By default, the SDK outputs five log files, `agorasdk.log`, `agorasdk_1.log`, `agorasdk_2.log`, `agorasdk_3.log`, `agorasdk_4.log`, each with a default size of 1024 KB.
+  /// These log files are encoded in UTF-8. The SDK writes the latest logs in `agorasdk.log`. When `agorasdk.log` is full, the SDK deletes the log file with the earliest modification
+  /// time among the other four, renames `agorasdk.log` to the name of the deleted log file, and creates a new `agorasdk.log` to record latest logs.
+  @JsonKey(includeIfNull: false)
+  LogConfig? logConfig;
+
+  /// Constructs a [RtcEngineConfig]
+  RtcEngineConfig(this.appId, {this.areaCode, this.logConfig});
+
+  /// @nodoc
+  factory RtcEngineConfig.fromJson(Map<String, dynamic> json) =>
+      _$RtcEngineConfigFromJson(json);
+
+  /// @nodoc
+  Map<String, dynamic> toJson() => _$RtcEngineConfigToJson(this);
 }
