@@ -23,12 +23,12 @@ class RtcSurfaceView extends StatefulWidget {
   /// - All uppercase English letters: A to Z.
   /// - All numeric characters: 0 to 9.
   /// - The space character.
-  /// - Punctuation characters and other symbols, including: "!", "#", "$", "%", "&", "(", ")", "+", "-", ":", ";", "<", "=", ".", ">", "?", "@", "[", "]", "^", "_", " {", "}", "|", "~", ",".
+  /// - Punctuation characters and other symbols, including: "!", "#", "$", "%", "&", "(", ")", "+", "-", ":", ";", "<", "=", ".", ">", "?", "@", "\[", "\]", "^", "_", " {", "}", "|", "~", ",".
   ///
   /// **Note**
   /// - The default value is the empty string "". Use the default value if the user joins the channel using the [RtcEngine.joinChannel] method in the [RtcEngine] class.
   /// - If the user joins the channel using the [RtcChannel.joinChannel] method in the [RtcChannel] class, set this parameter as the channelId of the [RtcChannel] object.
-  final String channelId;
+  final String? channelId;
 
   /// The rendering mode of the video view.
   final VideoRenderMode renderMode;
@@ -49,7 +49,7 @@ class RtcSurfaceView extends StatefulWidget {
   /// Callback signature for when a platform view was created.
   ///
   /// `id` is the platform view's unique identifier.
-  final PlatformViewCreatedCallback onPlatformViewCreated;
+  final PlatformViewCreatedCallback? onPlatformViewCreated;
 
   /// Which gestures should be consumed by the web view.
   ///
@@ -60,12 +60,12 @@ class RtcSurfaceView extends StatefulWidget {
   ///
   /// When this set is empty or null, the web view will only handle pointer events for gestures that
   /// were not claimed by any other gesture recognizer.
-  final Set<Factory<OneSequenceGestureRecognizer>> gestureRecognizers;
+  final Set<Factory<OneSequenceGestureRecognizer>>? gestureRecognizers;
 
   /// Constructs a [RtcSurfaceView]
   RtcSurfaceView({
-    Key key,
-    @required this.uid,
+    Key? key,
+    required this.uid,
     this.channelId,
     this.renderMode = VideoRenderMode.Hidden,
     this.mirrorMode = VideoMirrorMode.Auto,
@@ -82,9 +82,9 @@ class RtcSurfaceView extends StatefulWidget {
 }
 
 class _RtcSurfaceViewState extends State<RtcSurfaceView> {
-  int _id;
-  int _renderMode;
-  int _mirrorMode;
+  int? _id;
+  int? _renderMode;
+  int? _mirrorMode;
 
   @override
   Widget build(BuildContext context) {
@@ -163,32 +163,27 @@ class _RtcSurfaceViewState extends State<RtcSurfaceView> {
   }
 
   void setData() {
-    if (widget.uid == null) return;
     _channels[_id]?.invokeMethod('setData', {
       'data': {'uid': widget.uid, 'channelId': widget.channelId}
     });
   }
 
   void setRenderMode() {
-    if (widget.renderMode == null) return;
     _renderMode = VideoRenderModeConverter(widget.renderMode).value();
     _channels[_id]?.invokeMethod('setRenderMode', {'renderMode': _renderMode});
   }
 
   void setMirrorMode() {
-    if (widget.mirrorMode == null) return;
     _mirrorMode = VideoMirrorModeConverter(widget.mirrorMode).value();
     _channels[_id]?.invokeMethod('setMirrorMode', {'mirrorMode': _mirrorMode});
   }
 
   void setZOrderOnTop() {
-    if (widget.zOrderOnTop == null) return;
     _channels[_id]
         ?.invokeMethod('setZOrderOnTop', {'onTop': widget.zOrderOnTop});
   }
 
   void setZOrderMediaOverlay() {
-    if (widget.zOrderMediaOverlay == null) return;
     _channels[_id]?.invokeMethod(
         'setZOrderMediaOverlay', {'isMediaOverlay': widget.zOrderMediaOverlay});
   }
@@ -214,12 +209,12 @@ class RtcTextureView extends StatefulWidget {
   /// - All uppercase English letters: A to Z.
   /// - All numeric characters: 0 to 9.
   /// - The space character.
-  /// - Punctuation characters and other symbols, including: "!", "#", "$", "%", "&", "(", ")", "+", "-", ":", ";", "<", "=", ".", ">", "?", "@", "[", "]", "^", "_", " {", "}", "|", "~", ",".
+  /// - Punctuation characters and other symbols, including: "!", "#", "$", "%", "&", "(", ")", "+", "-", ":", ";", "<", "=", ".", ">", "?", "@", "\[", "\]", "^", "_", " {", "}", "|", "~", ",".
   ///
   /// **Note**
   /// - The default value is the empty string "". Use the default value if the user joins the channel using the [RtcEngine.joinChannel] method in the [RtcEngine] class.
   /// - If the user joins the channel using the [RtcChannel.joinChannel] method in the [RtcChannel] class, set this parameter as the channelId of the [RtcChannel] object.
-  final String channelId;
+  final String? channelId;
 
   /// The rendering mode of the video view.
   final VideoRenderMode renderMode;
@@ -230,7 +225,7 @@ class RtcTextureView extends StatefulWidget {
   /// Callback signature for when a platform view was created.
   ///
   /// `id` is the platform view's unique identifier.
-  final PlatformViewCreatedCallback onPlatformViewCreated;
+  final PlatformViewCreatedCallback? onPlatformViewCreated;
 
   /// Which gestures should be consumed by the web view.
   ///
@@ -241,12 +236,12 @@ class RtcTextureView extends StatefulWidget {
   ///
   /// When this set is empty or null, the web view will only handle pointer events for gestures that
   /// were not claimed by any other gesture recognizer.
-  final Set<Factory<OneSequenceGestureRecognizer>> gestureRecognizers;
+  final Set<Factory<OneSequenceGestureRecognizer>>? gestureRecognizers;
 
   /// Constructs a [RtcTextureView]
   RtcTextureView({
-    Key key,
-    @required this.uid,
+    Key? key,
+    required this.uid,
     this.channelId,
     this.renderMode = VideoRenderMode.Hidden,
     this.mirrorMode = VideoMirrorMode.Auto,
@@ -261,9 +256,9 @@ class RtcTextureView extends StatefulWidget {
 }
 
 class _RtcTextureViewState extends State<RtcTextureView> {
-  int _id;
-  int _renderMode;
-  int _mirrorMode;
+  int? _id;
+  int? _renderMode;
+  int? _mirrorMode;
 
   @override
   Widget build(BuildContext context) {
@@ -316,20 +311,17 @@ class _RtcTextureViewState extends State<RtcTextureView> {
   }
 
   void setData() {
-    if (widget.uid == null) return;
     _channels[_id]?.invokeMethod('setData', {
       'data': {'uid': widget.uid, 'channelId': widget.channelId}
     });
   }
 
   void setRenderMode() {
-    if (widget.renderMode == null) return;
     _renderMode = VideoRenderModeConverter(widget.renderMode).value();
     _channels[_id]?.invokeMethod('setRenderMode', {'renderMode': _renderMode});
   }
 
   void setMirrorMode() {
-    if (widget.mirrorMode == null) return;
     _mirrorMode = VideoMirrorModeConverter(widget.mirrorMode).value();
     _channels[_id]?.invokeMethod('setMirrorMode', {'mirrorMode': _mirrorMode});
   }
