@@ -1,17 +1,17 @@
 import 'package:agora_rtc_engine/rtc_engine.dart';
-import 'package:e2e/e2e.dart';
 import 'package:flutter_test/flutter_test.dart';
+import 'package:integration_test/integration_test.dart';
 
 void main() {
-  E2EWidgetsFlutterBinding.ensureInitialized();
+  IntegrationTestWidgetsFlutterBinding.ensureInitialized();
 
   group('basic', () {
-    RtcEngine engine;
+    RtcEngine? engine;
 
     // Connect to the Flutter driver before running any tests.
     setUpAll(() async {
-      engine = await RtcEngine.create('2b4b76e458cf439aa7cd313b9504f0a4');
-      engine.setEventHandler(RtcEngineEventHandler(warning: (warn) {
+      engine = await RtcEngine.create('');
+      engine!.setEventHandler(RtcEngineEventHandler(warning: (warn) {
         print('engine warn $warn');
       }, rtcStats: (stats) {
         print('engine rtcStats ${stats.toJson()}');
@@ -21,13 +21,13 @@ void main() {
     // Close the connection to the driver after the tests have completed.
     tearDownAll(() async {
       if (engine != null) {
-        await engine.destroy();
+        await engine!.destroy();
       }
     });
 
     testWidgets('joinChannel', (WidgetTester widgetTester) async {
-      await engine.joinChannel(null, '123', null, 0);
-      expect(await engine.getCallId(), isNotNull);
+      await engine!.joinChannel(null, '123', null, 0);
+      expect(await engine!.getCallId(), isNotNull);
     });
   });
 }
