@@ -4,6 +4,11 @@ const appId = 'd4813baec629491291f7b6ab4e06a9e3';
 // /// Please refer to https://docs.agora.io/en/Agora%20Platform/token
 const token = null;
 
-const channelId = 'captio_test';
-const uid = 0;
-// const stringUid = YOUR_STRING_UID;
+/// Your channel ID
+const channelId = YOUR_CHANNEL_ID;
+
+/// Your int user ID
+const uid = YOUR_UID;
+
+/// Your string user ID
+const stringUid = YOUR_STRING_UID;
