@@ -84,7 +84,10 @@ extension AgoraRtcRemoteAudioStats {
             "totalFrozenTime": totalFrozenTime,
             "frozenRate": frozenRate,
             "totalActiveTime": totalActiveTime,
-            "publishDuration": publishDuration
+            "publishDuration": publishDuration,
+            "qoeQuality": qoeQuality,
+            "qualityChangedReason": qualityChangedReason,
+            "mosValue": mosValue
         ]
     }
 }
@@ -105,7 +108,8 @@ extension AgoraRtcLocalVideoStats {
             "encodedFrameCount": encodedFrameCount,
             "codecType": codecType.rawValue,
             "txPacketLossRate": txPacketLossRate,
-            "captureFrameRate": captureFrameRate
+            "captureFrameRate": captureFrameRate,
+            "captureBrightnessLevel": captureBrightnessLevel.rawValue
         ]
     }
 }
