@@ -51,6 +51,7 @@ protocol RtcChannelAudioInterface {
 
     func muteAllRemoteAudioStreams(_ params: NSDictionary, _ callback: Callback)
 
+    @available(*, deprecated)
     func setDefaultMuteAllRemoteAudioStreams(_ params: NSDictionary, _ callback: Callback)
 }
 
@@ -59,7 +60,10 @@ protocol RtcChannelVideoInterface {
 
     func muteAllRemoteVideoStreams(_ params: NSDictionary, _ callback: Callback)
 
+    @available(*, deprecated)
     func setDefaultMuteAllRemoteVideoStreams(_ params: NSDictionary, _ callback: Callback)
+
+    func enableRemoteSuperResolution(_ params: NSDictionary, _ callback: Callback)
 }
 
 protocol RtcChannelVoicePositionInterface {
@@ -103,8 +107,10 @@ protocol RtcChannelMediaMetadataInterface {
 }
 
 protocol RtcChannelEncryptionInterface {
+    @available(*, deprecated)
     func setEncryptionSecret(_ params: NSDictionary, _ callback: Callback)
 
+    @available(*, deprecated)
     func setEncryptionMode(_ params: NSDictionary, _ callback: Callback)
 
     func enableEncryption(_ params: NSDictionary, _ callback: Callback)
@@ -163,15 +169,11 @@ class RtcChannelManager: NSObject, RtcChannelInterface {
     }
 
     @objc func destroy(_ params: NSDictionary, _ callback: Callback) {
-        var code: Int32? = -Int32(AgoraErrorCode.notInitialized.rawValue)
-        if let it = self[params["channelId"] as! String] {
-            code = rtcChannelMap.removeValue(forKey: it.getId()!)?.destroy()
-        }
-        callback.code(code)
+        callback.code(rtcChannelMap.removeValue(forKey:params["channelId"] as! String)?.destroy())
     }
 
     @objc func setClientRole(_ params: NSDictionary, _ callback: Callback) {
-        let role = AgoraClientRole(rawValue: params["role"] as! Int)!
+        let role = AgoraClientRole(rawValue: (params["role"] as! NSNumber).intValue)!
         if let options = params["options"] as? Dictionary<String, Any> {
             callback.code(self[params["channelId"] as! String]?.setClientRole(role, options: mapToClientRoleOptions(options)))
             return
@@ -180,7 +182,7 @@ class RtcChannelManager: NSObject, RtcChannelInterface {
     }
 
     @objc func joinChannel(_ params: NSDictionary, _ callback: Callback) {
-        callback.code(self[params["channelId"] as! String]?.join(byToken: params["token"] as? String, info: params["optionalInfo"] as? String, uid: params["optionalUid"] as! UInt, options: mapToChannelMediaOptions(params["options"] as! Dictionary)))
+        callback.code(self[params["channelId"] as! String]?.join(byToken: params["token"] as? String, info: params["optionalInfo"] as? String, uid: (params["optionalUid"] as! NSNumber).uintValue, options: mapToChannelMediaOptions(params["options"] as! Dictionary)))
     }
 
     @objc func joinChannelWithUserAccount(_ params: NSDictionary, _ callback: Callback) {
@@ -216,11 +218,11 @@ class RtcChannelManager: NSObject, RtcChannelInterface {
     }
 
     @objc func adjustUserPlaybackSignalVolume(_ params: NSDictionary, _ callback: Callback) {
-        callback.code(self[params["channelId"] as! String]?.adjustUserPlaybackSignalVolume(params["uid"] as! UInt, volume: params["volume"] as! Int32))
+        callback.code(self[params["channelId"] as! String]?.adjustUserPlaybackSignalVolume((params["uid"] as! NSNumber).uintValue, volume: (params["volume"] as! NSNumber).int32Value))
     }
 
     @objc func muteRemoteAudioStream(_ params: NSDictionary, _ callback: Callback) {
-        callback.code(self[params["channelId"] as! String]?.muteRemoteAudioStream(params["uid"] as! UInt, mute: params["muted"] as! Bool))
+        callback.code(self[params["channelId"] as! String]?.muteRemoteAudioStream((params["uid"] as! NSNumber).uintValue, mute: params["muted"] as! Bool))
     }
 
     @objc func muteAllRemoteAudioStreams(_ params: NSDictionary, _ callback: Callback) {
@@ -232,7 +234,7 @@ class RtcChannelManager: NSObject, RtcChannelInterface {
     }
 
     @objc func muteRemoteVideoStream(_ params: NSDictionary, _ callback: Callback) {
-        callback.code(self[params["channelId"] as! String]?.muteRemoteVideoStream(params["uid"] as! UInt, mute: params["muted"] as! Bool))
+        callback.code(self[params["channelId"] as! String]?.muteRemoteVideoStream((params["uid"] as! NSNumber).uintValue, mute: params["muted"] as! Bool))
     }
 
     @objc func muteAllRemoteVideoStreams(_ params: NSDictionary, _ callback: Callback) {
@@ -244,7 +246,7 @@ class RtcChannelManager: NSObject, RtcChannelInterface {
     }
 
     @objc func setRemoteVoicePosition(_ params: NSDictionary, _ callback: Callback) {
-        callback.code(self[params["channelId"] as! String]?.setRemoteVoicePosition(params["uid"] as! UInt, pan: params["pan"] as! Double, gain: params["gain"] as! Double))
+        callback.code(self[params["channelId"] as! String]?.setRemoteVoicePosition((params["uid"] as! NSNumber).uintValue, pan: (params["pan"] as! NSNumber).doubleValue, gain: (params["gain"] as! NSNumber).doubleValue))
     }
 
     @objc func setLiveTranscoding(_ params: NSDictionary, _ callback: Callback) {
@@ -272,45 +274,46 @@ class RtcChannelManager: NSObject, RtcChannelInterface {
     }
 
     @objc func setRemoteVideoStreamType(_ params: NSDictionary, _ callback: Callback) {
-        callback.code(self[params["channelId"] as! String]?.setRemoteVideoStream(params["uid"] as! UInt, type: AgoraVideoStreamType(rawValue: params["streamType"] as! Int)!))
+        callback.code(self[params["channelId"] as! String]?.setRemoteVideoStream((params["uid"] as! NSNumber).uintValue, type: AgoraVideoStreamType(rawValue: (params["streamType"] as! NSNumber).intValue)!))
     }
 
     @objc func setRemoteDefaultVideoStreamType(_ params: NSDictionary, _ callback: Callback) {
-        callback.code(self[params["channelId"] as! String]?.setRemoteDefaultVideoStreamType(AgoraVideoStreamType(rawValue: params["streamType"] as! Int)!))
+        callback.code(self[params["channelId"] as! String]?.setRemoteDefaultVideoStreamType(AgoraVideoStreamType(rawValue: (params["streamType"] as! NSNumber).intValue)!))
     }
 
     @objc func setRemoteUserPriority(_ params: NSDictionary, _ callback: Callback) {
-        callback.code(self[params["channelId"] as! String]?.setRemoteUserPriority(params["uid"] as! UInt, type: AgoraUserPriority(rawValue: params["userPriority"] as! Int)!))
+        callback.code(self[params["channelId"] as! String]?.setRemoteUserPriority((params["uid"] as! NSNumber).uintValue, type: AgoraUserPriority(rawValue: (params["userPriority"] as! NSNumber).intValue)!))
     }
 
     @objc func registerMediaMetadataObserver(_ params: NSDictionary, _ callback: Callback) {
-        var code = -AgoraErrorCode.notInitialized.rawValue
-        if let it = self[params["channelId"] as! String] {
-            let mediaObserver = MediaObserver { [weak self] in
-                self?.emitter(RtcEngineEvents.MetadataReceived, $0)
-            }
-            if it.setMediaMetadataDelegate(mediaObserver, with: .video) {
-                mediaObserverMap[it.getId()!] = mediaObserver
-                code = AgoraErrorCode.noError.rawValue
+        let channelId = params["channelId"] as! String
+        let mediaObserver = MediaObserver { [weak self] in
+            if var data = $0 {
+                data["channelId"] = channelId;
+                self?.emitter(RtcEngineEvents.MetadataReceived, data)
             }
         }
-        callback.code(Int32(code))
+        callback.resolve(self[channelId]) {
+            if $0.setMediaMetadataDelegate(mediaObserver, with: .video) {
+                mediaObserverMap[channelId] = mediaObserver
+            }
+            return nil
+        }
     }
 
     @objc func unregisterMediaMetadataObserver(_ params: NSDictionary, _ callback: Callback) {
-        var code = -AgoraErrorCode.notInitialized.rawValue
-        if let it = self[params["channelId"] as! String] {
-            if it.setMediaMetadataDelegate(nil, with: .video) {
-                mediaObserverMap.removeValue(forKey: it.getId()!)
-                code = AgoraErrorCode.noError.rawValue
+        let channelId = params["channelId"] as! String
+        callback.resolve(self[channelId]) {
+            if $0.setMediaMetadataDelegate(nil, with: .video) {
+                mediaObserverMap.removeValue(forKey: channelId)
             }
+            return nil
         }
-        callback.code(Int32(code))
     }
 
     @objc func setMaxMetadataSize(_ params: NSDictionary, _ callback: Callback) {
         callback.resolve(mediaObserverMap[params["channelId"] as! String]) {
-            $0.setMaxMetadataSize(params["size"] as! Int)
+            $0.setMaxMetadataSize((params["size"] as! NSNumber).intValue)
         }
     }
 
@@ -326,7 +329,7 @@ class RtcChannelManager: NSObject, RtcChannelInterface {
 
     @objc func setEncryptionMode(_ params: NSDictionary, _ callback: Callback) {
         var encryptionMode = ""
-        switch params["encryptionMode"] as! Int {
+        switch (params["encryptionMode"] as! NSNumber).intValue {
         case AgoraEncryptionMode.AES128XTS.rawValue:
             encryptionMode = "aes-128-xts"
         case AgoraEncryptionMode.AES128ECB.rawValue:
@@ -351,25 +354,20 @@ class RtcChannelManager: NSObject, RtcChannelInterface {
     }
 
     @objc func createDataStream(_ params: NSDictionary, _ callback: Callback) {
-        var code: Int32 = -Int32(AgoraErrorCode.notInitialized.rawValue)
+        let channel = self[params["channelId"] as! String]
         var streamId = 0
-        if let it = self[params["channelId"] as! String] {
-            code = it.createDataStream(&streamId, reliable: params["reliable"] as! Bool, ordered: params["ordered"] as! Bool)
+        if let config = params["config"] as? Dictionary<String, Any> {
+            callback.code(channel?.createDataStream(&streamId, config: mapToDataStreamConfig(config))) { _ in streamId }
+            return
         }
-        callback.code(code) { ignore in
-            streamId
-        }
+        callback.code(channel?.createDataStream(&streamId, reliable: params["reliable"] as! Bool, ordered: params["ordered"] as! Bool)) { _ in streamId }
     }
 
     @objc func sendStreamMessage(_ params: NSDictionary, _ callback: Callback) {
-        var code: Int32 = -Int32(AgoraErrorCode.notInitialized.rawValue)
-        if let it = self[params["channelId"] as! String] {
-            if let data = (params["message"] as! String).data(using: .utf8) {
-                code = it.sendStreamMessage(params["streamId"] as! Int, data: data)
-            } else {
-                code = -Int32(AgoraErrorCode.invalidArgument.rawValue)
-            }
-        }
-        callback.code(code)
+        callback.code(self[params["channelId"] as! String]?.sendStreamMessage((params["streamId"] as! NSNumber).intValue, data: (params["message"] as! String).data(using: .utf8)!))
+    }
+
+    @objc func enableRemoteSuperResolution(_ params: NSDictionary, _ callback: Callback) {
+        callback.code(self[params["channelId"] as! String]?.enableRemoteSuperResolution((params["uid"] as! NSNumber).uintValue, enabled: params["enable"] as! Bool))
     }
 }
