@@ -454,7 +454,7 @@ enum AudioScenario {
 
   /// Meeting scenario that mainly contains the human voice.
   ///
-  /// Since v3.2.0
+  /// Since v3.2.1
   @JsonValue(8)
   MEETING,
 }
@@ -538,7 +538,7 @@ enum AudioVoiceChanger {
   GENERAL_BEAUTY_VOICE_FEMALE_VITALITY,
 }
 
-/// The camera capturer configuration.
+/// The camera capture configuration.
 enum CameraCaptureOutputPreference {
   /// (default) Self-adapts the camera output parameters to the system performance and network conditions to balance CPU consumption and video preview quality.
   @JsonValue(0)
@@ -552,9 +552,9 @@ enum CameraCaptureOutputPreference {
   @JsonValue(2)
   Preview,
 
-  /// Internal use only.
+  /// Capture Dimensions determined by user.
   @JsonValue(3)
-  Unkown,
+  Manual,
 }
 
 /// The camera direction.
@@ -785,6 +785,10 @@ enum ConnectionChangedReason {
   /// - [ConnectionStateType.Reconnecting]
   @JsonValue(14)
   KeepAliveTimeout,
+
+  /// In cloud proxy mode, the proxy server connection interrupted.
+  @JsonValue(15)
+  ProxyServerInterrupted,
 }
 
 /// Connection states.
@@ -856,11 +860,17 @@ enum EncryptionMode {
   @JsonValue(3)
   AES256XTS,
 
-  /// 128-bit SM4 encryption, ECB mode.
-  ///
-  ///
+  /// @nodoc 128-bit SM4 encryption, ECB mode.
   @JsonValue(4)
   SM4128ECB,
+
+  /// 128-bit AES encryption, GCM mode.
+  @JsonValue(5)
+  AES128GCM,
+
+  /// 256-bit AES encryption, GCM mode.
+  @JsonValue(6)
+  AES256GCM,
 }
 
 /// Error codes occur when the SDK encounters an error that cannot be recovered automatically without any app intervention.
@@ -968,8 +978,6 @@ enum ErrorCode {
   InvalidChannelId,
 
   /// 103: Fails to get server resources in the specified region. Please try to specify another region.
-  ///
-  ///
   @JsonValue(103)
   NoServerResources,
 
@@ -1080,6 +1088,10 @@ enum ErrorCode {
   /// The format of the RTMP stream URL is not supported. Check whether the URL format is correct.
   @JsonValue(156)
   PublishStreamFormatNotSuppported,
+
+  /// The App lack necessary library file. Check whether the dynamic library is loaded.
+  @JsonValue(157)
+  ModuleNotFound,
 
   /// Fails to load the media engine.
   @JsonValue(1001)
@@ -1302,6 +1314,14 @@ enum LocalVideoStreamError {
   /// The local video encoding fails.
   @JsonValue(5)
   EncodeFailure,
+
+  /// (iOS only) The application is in the background.
+  @JsonValue(6)
+  CaptureInBackground,
+
+  /// (iOS only) The application is running in Slide Over, Split View, or Picture in Picture mode.
+  @JsonValue(7)
+  CaptureMultipleForegroundApps,
 }
 
 /// The state of the local video stream.
@@ -1872,8 +1892,6 @@ enum WarningCode {
   AdmInterruption,
 
   /// During a call, `AudioSessionCategory` should be set to `AVAudioSessionCategoryPlayAndRecord`, and the SDK monitors this value. If the `AudioSessionCategory` is set to other values, this warning code is triggered and the SDK will forcefully set it back to `AVAudioSessionCategoryPlayAndRecord`.
-  ///
-  ///
   @JsonValue(1029)
   AdmCategoryNotPlayAndRecord,
 
@@ -1893,14 +1911,10 @@ enum WarningCode {
   /// - Restart your audio device.
   /// - Restart your device where the app runs.
   /// - Upgrade the sound card drive.
-  ///
-  ///
   @JsonValue(1040)
   AdmNoDataReadyCallback,
 
   /// Audio device module: The audio recording device is different from the audio playback device, which may cause echoes problem. Agora recommends using the same audio device to record and playback audio.
-  ///
-  ///
   @JsonValue(1042)
   AdmInconsistentDevices,
 
@@ -2001,8 +2015,6 @@ enum StreamPublishState {
 }
 
 /// The subscribing state.
-///
-///
 enum StreamSubscribeState {
   /// The initial subscribing state after joining the channel.
   @JsonValue(0)
@@ -2064,9 +2076,11 @@ enum AudioEffectPreset {
   /// Turn off audio effects and use the original voice.
   @JsonValue(0x00000000)
   AudioEffectOff,
+
   /// An audio effect typical of a KTV venue.
   @JsonValue(0x02010100)
   RoomAcousticsKTV,
+
   /// An audio effect typical of a concert hall.
   ///
   /// **Note**
@@ -2074,6 +2088,7 @@ enum AudioEffectPreset {
   /// To achieve better audio effect quality, Agora recommends calling [RtcEngine.setAudioProfile] and setting the profile parameter to `MusicHighQuality(4)` or `MusicHighQualityStereo(5)` before setting this enumerator.
   @JsonValue(0x02010200)
   RoomAcousticsVocalConcert,
+
   /// An audio effect typical of a recording studio.
   ///
   /// **Note**
@@ -2081,6 +2096,7 @@ enum AudioEffectPreset {
   /// To achieve better audio effect quality, Agora recommends calling [RtcEngine.setAudioProfile] and setting the profile parameter to `MusicHighQuality(4)` or `MusicHighQualityStereo(5)` before setting this enumerator.
   @JsonValue(0x02010300)
   RoomAcousticsStudio,
+
   /// An audio effect typical of a vintage phonograph.
   ///
   /// **Note**
@@ -2088,6 +2104,7 @@ enum AudioEffectPreset {
   /// To achieve better audio effect quality, Agora recommends calling [RtcEngine.setAudioProfile] and setting the profile parameter to `MusicHighQuality(4)` or `MusicHighQualityStereo(5)` before setting this enumerator.
   @JsonValue(0x02010400)
   RoomAcousticsPhonograph,
+
   /// A virtual stereo effect that renders monophonic audio as stereo audio.
   ///
   /// **Note**
@@ -2095,6 +2112,7 @@ enum AudioEffectPreset {
   /// Call `setAudioProfile` and set the profile parameter to `MusicStandardStereo(3)` or `MusicHighQualityStereo(5)` before setting this enumerator; otherwise, the enumerator setting does not take effect.
   @JsonValue(0x02010500)
   RoomAcousticsVirtualStereo,
+
   /// A more spatial audio effect.
   ///
   /// **Note**
@@ -2102,6 +2120,7 @@ enum AudioEffectPreset {
   /// To achieve better audio effect quality, Agora recommends calling [RtcEngine.setAudioProfile] and setting the profile parameter to `MusicHighQuality(4)` or `MusicHighQualityStereo(5)` before setting this enumerator.
   @JsonValue(0x02010600)
   RoomAcousticsSpacial,
+
   /// A more ethereal audio effect.
   ///
   /// **Note**
@@ -2109,6 +2128,7 @@ enum AudioEffectPreset {
   /// To achieve better audio effect quality, Agora recommends calling [RtcEngine.setAudioProfile] and setting the profile parameter to `MusicHighQuality(4)` or `MusicHighQualityStereo(5)` before setting this enumerator.
   @JsonValue(0x02010700)
   RoomAcousticsEthereal,
+
   /// A 3D voice effect that makes the voice appear to be moving around the user. The default cycle period of the 3D voice effect is 10 seconds. To change the cycle period, call [RtcEngine.setAudioEffectParameters] after this method.
   ///
   /// **Note**
@@ -2117,6 +2137,7 @@ enum AudioEffectPreset {
   /// - If the 3D voice effect is enabled, users need to use stereo audio playback devices to hear the anticipated voice effect.
   @JsonValue(0x02010800)
   RoomAcoustics3DVoice,
+
   /// The voice of a middle-aged man.
   ///
   /// **Note**
@@ -2125,6 +2146,7 @@ enum AudioEffectPreset {
   /// - To achieve better audio effect quality, Agora recommends calling [RtcEngine.setAudioProfile] and setting the profile parameter to `MusicHighQuality(4)` or `MusicHighQualityStereo(5)` before setting this enumerator.
   @JsonValue(0x02020100)
   VoiceChangerEffectUncle,
+
   /// The voice of an old man.
   ///
   /// **Note**
@@ -2133,6 +2155,7 @@ enum AudioEffectPreset {
   /// - To achieve better audio effect quality, Agora recommends calling [RtcEngine.setAudioProfile] and setting the profile parameter to `MusicHighQuality(4)` or `MusicHighQualityStereo(5)` before setting this enumerator.
   @JsonValue(0x02020200)
   VoiceChangerEffectOldMan,
+
   /// The voice of a boy.
   ///
   /// **Note**
@@ -2141,6 +2164,7 @@ enum AudioEffectPreset {
   /// - To achieve better audio effect quality, Agora recommends calling [RtcEngine.setAudioProfile] and setting the profile parameter to `MusicHighQuality(4)` or `MusicHighQualityStereo(5)` before setting this enumerator.
   @JsonValue(0x02020300)
   VoiceChangerEffectBoy,
+
   /// The voice of a young woman.
   ///
   /// **Note**
@@ -2149,6 +2173,7 @@ enum AudioEffectPreset {
   /// - To achieve better audio effect quality, Agora recommends calling [RtcEngine.setAudioProfile] and setting the profile parameter to `MusicHighQuality(4)` or `MusicHighQualityStereo(5)` before setting this enumerator.
   @JsonValue(0x02020400)
   VoiceChangerEffectSister,
+
   /// The voice of a girl.
   ///
   /// **Note**
@@ -2157,6 +2182,7 @@ enum AudioEffectPreset {
   /// - To achieve better audio effect quality, Agora recommends calling [RtcEngine.setAudioProfile] and setting the profile parameter to `MusicHighQuality(4)` or `MusicHighQualityStereo(5)` before setting this enumerator.
   @JsonValue(0x02020500)
   VoiceChangerEffectGirl,
+
   /// The voice of Pig King, a character in Journey to the West who has a voice like a growling bear.
   ///
   /// **Note**
@@ -2164,6 +2190,7 @@ enum AudioEffectPreset {
   /// To achieve better audio effect quality, Agora recommends calling [RtcEngine.setAudioProfile] and setting the profile parameter to `MusicHighQuality(4)` or `MusicHighQualityStereo(5)` before setting this enumerator.
   @JsonValue(0x02020600)
   VoiceChangerEffectPigKing,
+
   /// The voice of Hulk.
   ///
   /// **Note**
@@ -2171,6 +2198,7 @@ enum AudioEffectPreset {
   /// To achieve better audio effect quality, Agora recommends calling [RtcEngine.setAudioProfile] and setting the profile parameter to `MusicHighQuality(4)` or `MusicHighQualityStereo(5)` before setting this enumerator.
   @JsonValue(0x02020700)
   VoiceChangerEffectHulk,
+
   /// An audio effect typical of R&B music.
   ///
   /// **Note**
@@ -2178,6 +2206,7 @@ enum AudioEffectPreset {
   /// Call [RtcEngine.setAudioProfile] and set the profile parameter to `MusicHighQuality(4)` or `MusicHighQualityStereo(5)` before setting this enumerator; otherwise, the enumerator setting does not take effect.
   @JsonValue(0x02030100)
   StyleTransformationRnB,
+
   /// An audio effect typical of popular music.
   ///
   /// **Note**
@@ -2185,6 +2214,7 @@ enum AudioEffectPreset {
   /// Call [RtcEngine.setAudioProfile] and set the profile parameter to `MusicHighQuality(4)` or `MusicHighQualityStereo(5)` before setting this enumerator; otherwise, the enumerator setting does not take effect.
   @JsonValue(0x02030200)
   StyleTransformationPopular,
+
   /// A pitch correction effect that corrects the user’s pitch based on the pitch of the natural C major scale. To change the basic mode and tonic pitch, call [RtcEngine.setAudioEffectParameters] after this method.
   ///
   /// **Note**
@@ -2199,6 +2229,7 @@ enum VoiceBeautifierPreset {
   /// Turn off voice beautifier effects and use the original voice.
   @JsonValue(0x00000000)
   VoiceBeautifierOff,
+
   /// A more magnetic voice.
   ///
   /// **Note**
@@ -2206,37 +2237,51 @@ enum VoiceBeautifierPreset {
   /// Agora recommends using this enumerator to process a male-sounding voice; otherwise, you may experience vocal distortion.
   @JsonValue(0x01010100)
   ChatBeautifierMagnetic,
+
   /// A fresher voice.
   ///
   /// Agora recommends using this enumerator to process a female-sounding voice; otherwise, you may experience vocal distortion.
   @JsonValue(0x01010200)
   ChatBeautifierFresh,
+
   /// A more vital voice.
   ///
   /// Agora recommends using this enumerator to process a female-sounding voice; otherwise, you may experience vocal distortion.
   @JsonValue(0x01010300)
   ChatBeautifierVitality,
+
+  ///  Singing beautifier effect.
+  @JsonValue(0x01020100)
+  SingingBeautifier,
+
   /// A more vigorous voice.
   @JsonValue(0x01030100)
   TimbreTransformationVigorous,
+
   /// A deeper voice.
   @JsonValue(0x01030200)
   TimbreTransformationDeep,
+
   /// A mellower voice.
   @JsonValue(0x01030300)
   TimbreTransformationMellow,
+
   /// A falsetto voice.
   @JsonValue(0x01030400)
   TimbreTransformationFalsetto,
+
   /// A fuller voice.
   @JsonValue(0x01030500)
   TimbreTransformationFull,
+
   /// A clearer voice.
   @JsonValue(0x01030600)
   TimbreTransformationClear,
+
   /// A more resounding voice.
   @JsonValue(0x01030700)
   TimbreTransformationResounding,
+
   /// A more ringing voice.
   @JsonValue(0x01030800)
   TimbreTransformationRinging,
@@ -2255,4 +2300,140 @@ enum AudienceLatencyLevelType {
   /// 2: (Default) Ultra low latency.
   @JsonValue(2)
   UltraLowLatency,
+}
+
+///  The output log level of the SDK.
+enum LogLevel {
+  /// Do not output any log.
+  @JsonValue(0x0000)
+  None,
+
+  /// (Default) Output logs of the `Fatal`, `Error`, `Warn`, and `Info` level.
+  /// We recommend setting your log filter as this level.
+  @JsonValue(0x0001)
+  Info,
+
+  /// Output logs of the `Fatal`, `Error`, and `Warn` level.
+  @JsonValue(0x0002)
+  Warn,
+
+  /// Output logs of the `Fatal` and `Error` level.
+  @JsonValue(0x0004)
+  Error,
+
+  /// Output logs of the `Fatal` level.
+  @JsonValue(0x0008)
+  Fatal,
+}
+
+/// The brightness level of the video image captured by the local camera.
+enum CaptureBrightnessLevelType {
+  /// The SDK does not detect the brightness level of the video image.
+  /// Wait a few seconds to get the brightness level from [CaptureBrightnessLevelType] in the next callback.
+  @JsonValue(-1)
+  Invalid,
+  /// The brightness level of the video image is normal.
+  @JsonValue(0)
+  Normal,
+  /// The brightness level of the video image is too bright.
+  @JsonValue(1)
+  Bright,
+  /// The brightness level of the video image is too dark.
+  @JsonValue(2)
+  Dark,
+}
+
+/// The reason why the super-resolution algorithm is not successfully enabled.
+enum SuperResolutionStateReason {
+  /// 0: The super-resolution algorithm is successfully enabled.
+  @JsonValue(0)
+  Success,
+
+  /// 1: The origin resolution of the remote video is beyond the range where the super-resolution algorithm can be applied.
+  @JsonValue(1)
+  StreamOverLimitation,
+
+  /// 2: Another user is already using the super-resolution algorithm.
+  @JsonValue(2)
+  UserCountOverLimitation,
+
+  /// 3: The device does not support the super-resolution algorithm.
+  @JsonValue(3)
+  DeviceNotSupported,
+}
+
+/// @nodoc
+enum UploadErrorReason {
+  @JsonValue(0)
+  Success,
+
+  @JsonValue(1)
+  NetError,
+
+  @JsonValue(2)
+  ServerError,
+}
+
+/// The cloud proxy type.
+enum CloudProxyType {
+  /// Do not use the cloud proxy.
+  @JsonValue(0)
+  None,
+
+  /// The cloud proxy for the UDP protocol.
+  @JsonValue(1)
+  UDP,
+
+  /// @nodoc The cloud proxy for the TCP protocol.
+  @JsonValue(2)
+  TCP,
+}
+
+/// Experience quality types.
+enum ExperienceQualityType {
+  /// 0: Good experience quality.
+  @JsonValue(0)
+  Good,
+
+  /// 1: Bad experience quality.
+  @JsonValue(1)
+  Bad,
+}
+
+/// The reason for poor QoE of the local user when receiving a remote audio stream.
+enum ExperiencePoorReason {
+  /// None, indicating good QoE of the local user.
+  @JsonValue(0)
+  None,
+  /// The remote user’s network quality is poor.
+  @JsonValue(1)
+  RemoteNetworkQualityPoor,
+  /// The local user’s network quality is poor.
+  @JsonValue(2)
+  LocalNetworkQualityPoor,
+  /// The local user’s Wi-Fi or mobile network signal is weak.
+  @JsonValue(4)
+  WirelessSignalPoor,
+  /// The local user enables both Wi-Fi and bluetooth, and their signals interfere with each other. As a result, audio transmission quality is undermined.
+  @JsonValue(8)
+  WifiBluetoothCoexist,
+}
+
+/// The options for SDK preset voice conversion effects.
+enum VoiceConversionPreset {
+  /// Turn off voice conversion effects and use the original voice.
+  @JsonValue(0)
+  Off,
+  /// A gender-neutral voice. To avoid audio distortion, ensure that you use this enumerator to process a female-sounding voice.
+  @JsonValue(50397440)
+  Neutral,
+  /// A sweet voice. To avoid audio distortion, ensure that you use this enumerator to process a female-sounding voice.
+  @JsonValue(50397696)
+  Sweet,
+  /// A steady voice. To avoid audio distortion, ensure that you use this enumerator to process a male-sounding voice.
+  @JsonValue(50397952)
+  Solid,
+  /// A deep voice. To avoid audio distortion, ensure that you use this enumerator to process a male-sounding voice.
+  @JsonValue(50398208)
+  Bass,
 }
