@@ -1,9 +1,10 @@
+import 'package:agora_rtc_engine/rtc_engine.dart';
+import 'package:flutter/material.dart';
 import 'package:flutter/services.dart';
 import 'package:flutter_test/flutter_test.dart';
-import 'package:agora_rtc_engine/rtc_engine.dart';
 
 void main() {
-  const MethodChannel channel = MethodChannel('agora_rtc_engine');
+  const channel = MethodChannel('agora_rtc_engine');
 
   TestWidgetsFlutterBinding.ensureInitialized();
 
@@ -11,6 +12,12 @@ void main() {
     channel.setMockMethodCallHandler((MethodCall methodCall) async {
       return '42';
     });
+  });
+
+  test('LiveTranscoding', () {
+    var transcoding = LiveTranscoding(null, backgroundColor: Colors.red);
+    print(transcoding.backgroundColor);
+    print(transcoding.toJson());
   });
 
   tearDown(() {
